feat: feat: 서버 실행 시 csv에서 추출한 상품 연관 정보를 SpringJdbcTemplate을 이용하여 벌크 삽입 (상품 등록 -> 전체 상품 조회 -> 해당 정보를 이용하여 상품 연관 정보 등록)

+++ src/main/java/com/example/techlabs/csv/ProductRelationshipCsvBean.java
@@ -21,5 +21,5 @@ public class ProductRelationshipCsvBean extends CsvBean {
     private BigDecimal score;
 
     @CsvBindByPosition(position = 3)
-    private BigDecimal rank;
+    private Long rank;
 }
